backend/requirements.txt
django==3.2
flake8==3.8.4  # https://github.com/PyCQA/flake8
black==20.8b1  # https://github.com/ambv/black
django-environ==0.4.5  # https://github.com/joke2k/django-environ
psycopg2-binary==2.8.6  # https://github.com/psycopg/psycopg2
django-extensions==3.0.9  # https://github.com/django-extensions/django-extensions
channels==3.0.4
channels-redis
djangorestframework==3.12.4
markdown==3.3.4
django-filter==2.4.0